Move redacted_url into crate and stop using in uv-auth
