+++ crates/uv-auth/src/credentials.rs
@@ -321,26 +321,6 @@ impl Credentials {
     }
 }
 
-/// Return a version of the URL with redacted credentials, allowing the generic `git` username (without a password)
-/// in SSH URLs, as in, `ssh://git@github.com/...`.
-pub fn redacted_url(url: &Url) -> Cow<'_, Url> {
-    let no_credentials = url.username().is_empty() && url.password().is_none();
-    let generic_git_username =
-        url.scheme() == "ssh" && url.username() == "git" && url.password().is_none();
-    if no_credentials || generic_git_username {
-        Cow::Borrowed(url)
-    } else {
-        let mut url = url.clone();
-        if url.password().is_some() {
-            let _ = url.set_password(Some("****"));
-        // A username on its own might be a secret token.
-        } else if url.username() != "" {
-            let _ = url.set_username("****");
-        }
-        Cow::Owned(url)
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use insta::assert_debug_snapshot;
