+++ crates/uv-client/Cargo.toml
@@ -24,6 +24,7 @@ uv-pep508 = { workspace = true }
 uv-platform-tags = { workspace = true }
 uv-pypi-types = { workspace = true }
 uv-small-str = { workspace = true }
+uv-redacted = { workspace = true }
 uv-static = { workspace = true }
 uv-torch = { workspace = true }
 uv-version = { workspace = true }
