+++ crates/uv-git/src/credentials.rs
@@ -2,8 +2,9 @@ use std::collections::HashMap;
 use std::sync::{Arc, LazyLock, RwLock};
 use tracing::trace;
 use url::Url;
-use uv_auth::{redacted_url, Credentials};
+use uv_auth::Credentials;
 use uv_cache_key::RepositoryUrl;
+use uv_redacted::redacted_url;
 
 /// Global authentication cache for a uv invocation.
 ///
